+++ tests/utils/simplified_test_function.py
@@ -132,6 +132,7 @@ def simplified_test_function(test_func):
                 if exp_exc_types:
                     with pytest.raises(exp_exc_types):
                         if condition == 'pdb':
+                            # pylint: disable=forgotten-debug-statement
                             pdb.set_trace()
 
                         test_func(testcase, **kwargs)  # expecting an exception
@@ -142,6 +143,7 @@ def simplified_test_function(test_func):
                     # exception).
                 else:
                     if condition == 'pdb':
+                        # pylint: disable=forgotten-debug-statement
                         pdb.set_trace()
 
                     test_func(testcase, **kwargs)  # not expecting an exception
@@ -153,6 +155,7 @@ def simplified_test_function(test_func):
                 if exp_exc_types:
                     with pytest.raises(exp_exc_types):
                         if condition == 'pdb':
+                            # pylint: disable=forgotten-debug-statement
                             pdb.set_trace()
 
                         test_func(testcase, **kwargs)  # expecting an exception
@@ -160,6 +163,7 @@ def simplified_test_function(test_func):
                     ret = None  # Debugging hint
                 else:
                     if condition == 'pdb':
+                        # pylint: disable=forgotten-debug-statement
                         pdb.set_trace()
 
                     test_func(testcase, **kwargs)  # not expecting an exception