+++ docs/changes.rst
@@ -19,7 +19,7 @@ Released: not yet
 * Mitigated the coveralls HTTP status 422 by pinning coveralls-python to
   <3.0.0 (issue #55).
 
-* Fixed issues raised by new Pylint 2.9.
+* Fixed issues raised by new Pylint 2.9 and 2.10.
 
 * Fixed a dependency error that caused importlib-metadata to be installed on
   Python 3.8, while it is included in the Python base.